Replaced redcal module with redcal2.
