+++ hera_gpu/__init__.py
@@ -2,4 +2,3 @@ name = "hera_gpu"
 from .median_filter import MedianFilter
 from .vis import vis_gpu
 from .redcal import omnical
-from .redcal2 import omnical as omnical2
